+++ apps/admin-panel/src/contexts/index.ts
@@ -1,1 +1,3 @@
+export { ApiContext } from "@contexts/api-context";
+export { AuthContext } from "@contexts/auth-context";
 export { ThemeManagerContext } from "@contexts/theme-manager-context";