+++ apps/admin-panel/src/hooks/index.ts
@@ -1,1 +1,3 @@
+export { useApi } from "@hooks/useApi";
+export { useAuth } from "@hooks/useAuth";
 export { useThemeManager } from "@hooks/useThemeManager";