+++ apps/admin-panel/src/providers/AppProviders.tsx
@@ -1,14 +1,17 @@
+import { ApiProvider } from "@providers/ApiProvider";
 import { AppThemeProvider } from "@providers/AppThemeProvider";
-import { QueryProvider } from "@providers/QueryProvider";
+import { AuthProvider } from "@providers/AuthProvider";
 import { ThemeManagerProvider } from "@providers/ThemeManagerProvider";
 import type { PropsWithChildren, ReactElement } from "react";
 
 export const AppProviders = ({ children }: PropsWithChildren): ReactElement => {
   return (
     <ThemeManagerProvider>
-      <QueryProvider>
-        <AppThemeProvider>{children}</AppThemeProvider>
-      </QueryProvider>
+      <AppThemeProvider>
+        <ApiProvider>
+          <AuthProvider>{children}</AuthProvider>
+        </ApiProvider>
+      </AppThemeProvider>
     </ThemeManagerProvider>
   );
 };