+++ apps/admin-panel/src/routes/router.tsx
@@ -1,9 +1,13 @@
 import { DashboardPage } from "@pages/DashboardPage";
-import { createBrowserRouter } from "react-router-dom";
+import { createBrowserRouter, Navigate } from "react-router-dom";
 
 export const router = createBrowserRouter([
   {
     element: <DashboardPage />,
     path: "/",
+  },
+  {
+    element: <Navigate replace to="/" />,
+    path: "*",
   },
 ]);